move functions + minor tweaks
